Tidy up code formatting and includes
